修改工时统计BUG，添加搜索条件，在项目列表时搜索项目工作，则展示项目工作，搜索部门工作，则展示类型为2的部门工作

+++ app/src/main/java/com/aorise/weeklyreport/bean/PlanWorkTimeSettingBean.java
@@ -47,7 +47,7 @@ public class PlanWorkTimeSettingBean extends BaseObservable implements Serializa
      */
 
     private float id;
-    private float year;
+    private int year;
     private float january;
     private float february;
     private float march;
@@ -73,7 +73,7 @@ public class PlanWorkTimeSettingBean extends BaseObservable implements Serializa
         return year;
     }
 
-    public void setYear(float year) {
+    public void setYear(int year) {
         this.year = year;
     }
